@@ -1,5 +1,5 @@
 [project]
-name = "api"
+name = "synthgenie"
 version = "0.1.0"
 description = "AI backend for synthgenie"
 readme = "README.md"
@@ -17,19 +17,19 @@ classifiers = [
 authors = [{ name = "zerubeus", email = "zerubeus@gmail.com" }]
 
 dependencies = [
-    "fastapi>=0.115.12",
-    "pydantic-ai[logfire]>=0.0.46",
+    "fastapi>=0.116.1",
+    "pydantic-ai[logfire]>=0.4.4",
     "python-dotenv>=1.1.0",
     "sentry-sdk[fastapi]>=2.25.0",
-    "psycopg2-binary>=2.9.9",
+    "psycopg2-binary>=2.9.10",
 ]
 
 [dependency-groups]
-dev = ["black>=25.1.0", "colorlog>=6.9.0", "pyright>=1.1.390", "ruff>=0.6.9"]
+dev = ["colorlog>=6.9.0", "pyright>=1.1.403", "ruff>=0.12.4"]
 
 [tool.ruff]
 line-length = 120
-target-version = "py39"
+target-version = "py312"
 include = ["synthgenie/**/*.py"]
 
 [tool.ruff.lint]
@@ -44,9 +44,6 @@ extend-select = [
     "TID251", # Bans certain types of relative imports for consistent import style (flake8-tidy-imports)
 ]
 flake8-quotes = { inline-quotes = "single", multiline-quotes = "double" }
-isort = { combine-as-imports = true, known-first-party = [
-    "synthgenie",
-] } # Sort imports
 mccabe = { max-complexity = 100 } # Check McCabe complexity
 ignore = [
     "D100", # ignore missing docstring in module
@@ -55,6 +52,10 @@ ignore = [
     "D105", # ignore missing docstring in magic methods
     "D107", # ignore missing docstring in __init__ methods
 ]
+
+[tool.ruff.lint.isort]
+combine-as-imports = true
+known-first-party = ["synthgenie"]
 
 [tool.ruff.lint.pydocstyle]
 convention = "google"
@@ -76,7 +77,8 @@ reportUnnecessaryIsInstance = false
 reportUnnecessaryTypeIgnoreComment = true
 reportMissingModuleSource = false
 include = ["synthgenie"]
-venvPath = ".venv"
+venvPath = "."
+venv = ".venv"
 
 [project.scripts]
 start = "synthgenie.app:app"
