@@ -1,5 +1,4 @@
 from datetime import datetime
-from typing import Optional
 
 from pydantic import BaseModel
 
@@ -20,4 +19,4 @@ class RevokeRequest(BaseModel):
 class ApiKeyUsage(BaseModel):
     key: str
     request_count: int
-    last_used_at: Optional[datetime] = None
+    last_used_at: datetime | None = None
