@@ -5,7 +5,7 @@ These parameters are controlled via NRPN messages.
 
 from pydantic_ai import RunContext
 
-from synthgenie.schemas.agent import SynthGenieResponse
+from synthgenie.synthesizers.shared.schemas.agent import SynthGenieResponse
 
 
 def set_arp_rate(ctx: RunContext, value: int, midi_channel: int = 3) -> SynthGenieResponse:
